+++ NWN.Anvil/src/main/API/Events/Game/TriggerEvents/TriggerEvents.OnDisarmed.cs
@@ -1,0 +1,44 @@
+using System;
+using Anvil.API.Events;
+using NWN.Core;
+
+namespace Anvil.API.Events
+{
+  /// <summary>
+  /// Events for drawn, world-placed triggers.
+  /// </summary>
+  public static partial class TriggerEvents
+  {
+    [GameEvent(EventScriptType.TriggerOnDisarmed)]
+    public sealed class OnDisarmed : IEvent
+    {
+      /// <summary>
+      /// Gets the <see cref="NwTrigger"/> that was disarmed.
+      /// </summary>
+      public NwTrigger Trigger { get; } = NWScript.OBJECT_SELF.ToNwObject<NwTrigger>();
+
+      /// <summary>
+      /// Gets the <see cref="NwCreature"/> who disarmed this trigger.
+      /// </summary>
+      public NwCreature DisarmedBy { get; } = NWScript.GetLastDisarmed().ToNwObject<NwCreature>();
+
+      NwObject IEvent.Context
+      {
+        get => Trigger;
+      }
+    }
+  }
+}
+
+namespace Anvil.API
+{
+  public sealed partial class NwTrigger
+  {
+    /// <inheritdoc cref="TriggerEvents.OnDisarmed"/>
+    public event Action<TriggerEvents.OnDisarmed> OnDisarmed
+    {
+      add => EventService.Subscribe<TriggerEvents.OnDisarmed, GameEventFactory, GameEventFactory.RegistrationData>(this, new GameEventFactory.RegistrationData(this), value);
+      remove => EventService.Unsubscribe<TriggerEvents.OnDisarmed, GameEventFactory>(this, value);
+    }
+  }
+}
