+++ NWN.Anvil/src/main/API/Events/Game/TriggerEvents/TriggerEvents.OnExit.cs
@@ -1,0 +1,44 @@
+using System;
+using Anvil.API.Events;
+using NWN.Core;
+
+namespace Anvil.API.Events
+{
+  /// <summary>
+  /// Events for drawn, world-placed triggers.
+  /// </summary>
+  public static partial class TriggerEvents
+  {
+    [GameEvent(EventScriptType.TriggerOnObjectExit)]
+    public sealed class OnExit : IEvent
+    {
+      /// <summary>
+      /// Gets the <see cref="NwTrigger"/> that was left.
+      /// </summary>
+      public NwTrigger Trigger { get; } = NWScript.OBJECT_SELF.ToNwObject<NwTrigger>();
+
+      /// <summary>
+      /// Gets the <see cref="NwGameObject"/> that left this <see cref="NwTrigger"/>.
+      /// </summary>
+      public NwGameObject ExitingObject { get; } = NWScript.GetExitingObject().ToNwObject<NwGameObject>();
+
+      NwObject IEvent.Context
+      {
+        get => Trigger;
+      }
+    }
+  }
+}
+
+namespace Anvil.API
+{
+  public sealed partial class NwTrigger
+  {
+    /// <inheritdoc cref="TriggerEvents.OnExit"/>
+    public event Action<TriggerEvents.OnExit> OnExit
+    {
+      add => EventService.Subscribe<TriggerEvents.OnExit, GameEventFactory, GameEventFactory.RegistrationData>(this, new GameEventFactory.RegistrationData(this), value);
+      remove => EventService.Unsubscribe<TriggerEvents.OnExit, GameEventFactory>(this, value);
+    }
+  }
+}
